+++ swing/common-ui/src/main/java/module-info.java
@@ -64,6 +64,7 @@ module is.codion.swing.common.ui {
 	exports is.codion.swing.common.ui.border;
 	exports is.codion.swing.common.ui.color;
 	exports is.codion.swing.common.ui.control;
+	exports is.codion.swing.common.ui.cursor;
 	exports is.codion.swing.common.ui.dialog;
 	exports is.codion.swing.common.ui.frame;
 	exports is.codion.swing.common.ui.icon;